pyproject.toml
[tool.poetry]
name = "hass-traeger"
version = "0.1.0"
description = ""
authors = ["captivus <366332+captivus@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
aiohttp = "3.9.5"
async-timeout = "4.0.3"
paho-mqtt = "2.1.0"
requests = "2.32.2"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pipreqs = "^0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
